% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_long.R
\docType{data}
\name{dummy_long}
\alias{dummy_long}
\title{SDC Long Data}
\format{
A data frame with 27 rows and 3 columns:
\describe{
\item{line}{Test groups containing counts}
\item{age_group}{Age category for counts}
\item{value}{Counts for each age category}
}
}
\source{
Dummy data created for the purpose of understanding the SDC App.
}
\usage{
dummy_long
}
\description{
Test Data in a Long Format for SDC App
}
\keyword{datasets}
