% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_wide.R
\docType{data}
\name{dummy_wide}
\alias{dummy_wide}
\title{SDC Wide Data}
\format{
A data frame with 27 rows and 3 columns:
\describe{
\item{line}{Test groups containing counts}
\item{20-24}{Age counts of 20-24 yrs old}
\item{25-29}{Age counts of 25-29 yrs old}
\item{30-34}{Age counts of 30-34 yrs old}
\item{35-39}{Age counts of 35-39 yrs old}
\item{40-44}{Age counts of 40-44 yrs old}
\item{45-49}{Age counts of 45-49 yrs old}
\item{50-54}{Age counts of 50-54 yrs old}
\item{55-59}{Age counts of 55-59 yrs old}
\item{Total}{Sum of all prior age counts}
}
}
\source{
Dummy data created for the purpose of understanding the SDC App.
}
\usage{
dummy_wide
}
\description{
Test Data in a Wide Format for SDC App. Same information as SDC Long Data but in a different format
}
\keyword{datasets}
