% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions.R
\name{SelectBox_Update}
\alias{SelectBox_Update}
\title{SelectBox_Update}
\usage{
SelectBox_Update(SDC_data)
}
\arguments{
\item{SDC_data}{input data containing variable names}
}
\value{
cb_options Variables names given in select box
}
\description{
Used for generate variable names in selectboxes. Can only be used inside Shiny App
}
