% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disclosure_functions.R
\name{Stat_Primary_Supress}
\alias{Stat_Primary_Supress}
\title{Perform Primary Suppression on Selected Variables}
\usage{
Stat_Primary_Supress(
  orig_data,
  var_choice,
  char_supp = "*",
  sup_cond,
  zero = TRUE
)
}
\arguments{
\item{orig_data}{A data frame or tibble containing the input data to have primary suppression applied.}

\item{var_choice}{A character vector specifying the names of the variables to be rounded. All specified variables must exist in orig_data.}

\item{char_supp}{A character used to indicate suppression. Default is \code{*}. Must be either \code{*} or \code{c}.}

\item{sup_cond}{A single positive integer. Values equal to or below this threshold will be suppressed.}

\item{zero}{A single logical value indicating whether zero should be suppressed. Default is TRUE. If TRUE, zero values will remain un-suppressed.}
}
\value{
A data frame or tibble with primary suppression applied to the specified variables. Original NA values are preserved.
}
\description{
This function applies primary suppression to specified variables within a dataset. Values equal to or below a given threshold are replaced with a suppression character.
}
\examples{
inp_data <- dummy_wide
ps_vars <- c("20-24", "25-29", "30-34", "35-39", "40-44", "45-49", "50-54", "55-59", "Total")
ps_char <- "*"
ps_cond <- 3
z_cond <- TRUE
ps_data <- Stat_Primary_Supress(inp_data, ps_vars, ps_char, ps_cond, z_cond)
inp_data
ps_data
}
