% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disclosure_functions.R
\name{Stat_Round}
\alias{Stat_Round}
\title{Round Selected Variables in a Dataset}
\usage{
Stat_Round(orig_data, var_choice, round_cond)
}
\arguments{
\item{orig_data}{A data frame containing the input data to be rounded.}

\item{var_choice}{A character vector specifying the names of the variables to be rounded.}

\item{round_cond}{A single positive whole integer indicating the rounding base.}
}
\value{
A data frame with the selected variables rounded to the specified base.
}
\description{
This function rounds specified variables within a dataset to a chosen base. It handles missing values by temporarily replacing them with a high value, which is reverted back to \code{NA} after rounding.
}
\examples{
inp_data <- dummy_wide
r_vars <- c("20-24", "25-29", "30-34", "35-39", "40-44", "45-49", "50-54", "55-59", "Total")
r_base <- 3
r_data <- Stat_Round(inp_data, r_vars, r_base)
inp_data
r_data
}
