% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disclosure_functions.R
\name{Stat_Secondary_Supress}
\alias{Stat_Secondary_Supress}
\title{Stat_Secondary_Supress Function}
\usage{
Stat_Secondary_Supress(
  orig_data,
  pri_var_choice,
  sec_var_choice,
  char_supp,
  sup_cond,
  zero
)
}
\arguments{
\item{orig_data}{Input Data to be disclosed}

\item{pri_var_choice}{Choice of variables to be suppressed using Primary Suppression}

\item{sec_var_choice}{Choice of variables to be suppressed using Secondary Suppression}

\item{char_supp}{Character used to indicate suppression}

\item{sup_cond}{Values equal to or below this value will be suppressed}

\item{zero}{Indicates if 0 should be suppressed}
}
\value{
secondary_data Data with Primary & Secondary Suppression Applied
}
\description{
Performs Primary & Secondary Suppression on selected variables within a Dataset
}
\examples{
inp_data <- dummy_wide
ps_vars <- "Total"
ss_vars <- c("20-24", "25-29", "30-34", "35-39", "40-44", "45-49", "50-54", "55-59")
s_char <- "*"
s_cond <- 3
z_cond <- TRUE
s_data <- Stat_Secondary_Supress(inp_data, ps_vars, ss_vars, s_char, s_cond, z_cond)
inp_data
s_data
}
