% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disclosure_functions.R
\name{Stat_Swap}
\alias{Stat_Swap}
\title{Stat_Swap Function}
\usage{
Stat_Swap(orig_data, var_choice, swap_cond)
}
\arguments{
\item{orig_data}{Input Data to be disclosed}

\item{var_choice}{Choice of variables to be swapped}

\item{swap_cond}{Swapping condtion meaning values below this in a dataset must be swapped}
}
\value{
Disclosed Data with Swapping Applied
}
\description{
Performs rounding on selected variables within a Dataset
}
\examples{
inp_data <- dummy_wide
s_vars <- c("20-24", "25-29", "30-34", "35-39", "40-44", "45-49", "50-54", "55-59", "Total")
s_cond <- 3
s_data <- Stat_Swap(inp_data, s_vars, s_cond)
inp_data
s_data
}
