% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions.R
\name{Table_Render}
\alias{Table_Render}
\title{Table_Render}
\usage{
Table_Render(dataset, cb)
}
\arguments{
\item{dataset}{Dataset to be rendered into table}

\item{cb}{Allows for interactive visualisation from Table}
}
\value{
output_data Visualised Table
}
\description{
Used for rendering tables within the SDC App. Can only be used inside Shiny App
}
