% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disclosure_functions.R
\name{Stat_Primary_Supress}
\alias{Stat_Primary_Supress}
\title{Stat_Primary_Supress Function}
\usage{
Stat_Primary_Supress(orig_data, var_choice, char_supp, sup_cond, zero)
}
\arguments{
\item{orig_data}{Input Data to be disclosed}

\item{var_choice}{Choice of variables to be suppressed}

\item{char_supp}{Character used to indicate suppression}

\item{sup_cond}{Values equal to or below this value will be suppressed}

\item{zero}{Indicates if 0 should be suppressed}
}
\value{
Disclosed Data with Primary Suppression Applied
}
\description{
Performs Primary Suppression on selected variables within a Dataset
}
\examples{
inp_data <- dummy_wide
ps_vars <- c("20-24", "25-29", "30-34", "35-39", "40-44", "45-49", "50-54", "55-59", "Total")
ps_char <- "*"
ps_cond <- "3"
z_cond <- TRUE
ps_data <- Stat_Primary_Supress(inp_data, ps_vars, ps_char, ps_cond, z_cond)
inp_data
ps_data
}
