% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disclosure_functions.R
\name{Stat_Round}
\alias{Stat_Round}
\title{Stat_Round Function}
\usage{
Stat_Round(orig_data, var_choice, round_cond)
}
\arguments{
\item{orig_data}{Input Data to be disclosed}

\item{var_choice}{Choice of variables to be rounded}

\item{round_cond}{Rounding Base}
}
\value{
rounded_data Disclosed Data with Rounding Applied
}
\description{
Performs rounding on selected variables within a dataset to a chosen base
}
\examples{
inp_data <- dummy_wide
r_vars <- c("20-24", "25-29", "30-34", "35-39", "40-44", "45-49", "50-54", "55-59", "Total")
r_base <- 3
r_data <- Stat_Round(inp_data, r_vars, r_base)
inp_data
r_data
}
