% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disclosure_functions.R
\name{Stat_Swap}
\alias{Stat_Swap}
\title{Swap Selected Variables in a Dataset}
\usage{
Stat_Swap(orig_data, var_choice, swap_cond)
}
\arguments{
\item{orig_data}{A dataframe or tibble containing the input data to be disclosed. This dataset
should include the variables that need to be swapped.}

\item{var_choice}{A character vector specifying the names of the variables to be swapped. These
variables must exist in the \code{orig_data}.}

\item{swap_cond}{A single positive whole integer indicating the swapping condition. Values
below this threshold in the specified variables will be swapped.}
}
\value{
A dataframe or tibble with the swapping applied to the selected variables.
The original data is returned if no numeric variables are selected.
}
\description{
Performs swapping on selected variables within a dataset by swapping values
that are below a specified condition.
}
\examples{
inp_data <- dummy_wide
s_vars <- c("20-24", "25-29", "30-34", "35-39", "40-44", "45-49", "50-54", "55-59", "Total")
s_cond <- 5
s_data <- Stat_Swap(inp_data, s_vars, s_cond)
inp_data
s_data
}
